Rename variables and files
