+++ parser/main.py
@@ -4,7 +4,7 @@ from crawler_current_events import get_text, parse_13f_filing_detail_urls, parse
     parse_primary_doc_xml_and_infotable_xml_urls, parse_primary_doc_xml_url, parse_infotable_xml_url
 from database_connection import session
 from data_13f import data_13f_row
-from filing_13f import parse_primary_doc_root, parse_primary_doc_cik, \
+from parsing_13f_filing import parse_primary_doc_root, parse_primary_doc_cik, \
     parse_primary_doc_company_name, parse_primary_doc_accepted_filing_date
 from models import EdgarFiling, Company
